added many tests, bug fixes
`normalizePath` now called at correct point in `normalizeArgs`.

added `requiredArgsCount` argument to `normalizeArgs`. used if
there are some arguments that are not required to ensure path
isn't added if not needed. only used in `remove`.

Fixed `push`, `get` so they work with a specified path argument.

Added a bunch of tests for performing actions with a specified
path argument. added tests for dotted paths.

There is still a bug in `move` when called with a path
(i've disabled the list move test for now). The test throws a
"TypeError: object is not a function". I cannot figure out why.
I believe there is a typo in `move` at the line

    if(cb) cb.apply(cb, arguments);

which I believe should be

    if(cb) cb.apply(this, arguments);

but changing it doesn't fix the problem, and I wanted to be sure
it is incorrect before changing it.