mongo.coffee won't bail out if DB was already open
If you use mongo.coffee with the client option (passing in a mongodb.Db instance), mongo.coffee won't open it if it was already open. Before, it would bail out of the Db was already open. Problematic, because mongodb.MongoClient.connect provides an already-opened DB instance (at least in mongodb 1.4.30), so it's quite likely that mongo.coffee will receive an already-open Db.