Make context.push() not break the context
Previously, context.push() was altering the context's path to include the new item. Consequently, the path would point to a different place to where callers expected. This also meant it worked exactly once, and any subsequent operations would be incorrect or even fail (pushes fail).

By cloning the path first, the context is maintained as pointing to the same place.